Load a simple controller
RestController class is a simple example of how to use rest controllers in a spring app. We use a couple of annotations like "RestController" and "GetMapping" to indicate that we are using a GET method. We can use other annotations for the rest of HTTP methods from a RESTful API like POST and DELETE, among other ones.